ci: speed up E2E and server checks
E2E (ProtoFleet E2E Tests):
- Replace the per-file matrix (~46 jobs) with 8 round-robin shards per
  project (16 jobs). Round-robin by file index breaks the alphabetical
  adjacency that Playwright's native --shard would cluster, so heavy
  suites spread across shards instead of landing together.
- Split racks.spec.ts (10 serial tests, ~9 min and the suite long pole)
  into four focused spec files sharing one hooks/constants module
  (helpers/racksTestSetup.ts). With round-robin the four pieces land on
  four separate shards, dropping the long pole from ~10 min to ~6 min.
  workers/fullyParallel are unchanged, so tests still run serially within
  a shard (one backend per shard): no new concurrency, no new flakiness.
- Cache ~/.npm instead of a built node_modules (which corrupts native
  optional deps), and skip `playwright install --with-deps` on a browser
  cache hit.

Server Checks:
- Split the single serial job (fmt -> lint -> build -> test) into parallel
  lint and build+test jobs sharing the Go cache, roughly halving wall-clock.
